Move function definitions from LinearAlgebraUtils.h into LinearAlgebraUtils.cpp
Out of probably laziness, we had not split LinearAlgebraUtils.h into a
header and a `.cpp` file. Instead, all the functions were defined as
static inline.

This is not a good practice in general, but also, it disallows to use
generated headers that should not be included in headers, such as
`ATen/cuda/CUDAConfig.h`, which states:

> // NB: This header MUST NOT be included from other headers; it should
> // only be included from C++ files.

~~This PR moves the definition of the functions into `LinearAlgebraUtils.cpp` to be able to include in the `.cpp` file the `CUDAConfig.h` header safely in a later PR in this stack.~~

It so happens that I did not need to use the `CUDAConfig.h` file in the
next PR of the stack. As such, this PR is not strictily necessary, but
I think it's worth it regardless.

I needed to sprinkle a few TORCH_API decorators on the functions that
are used both on cpu and cuda builds. If you ever encounter a linker
error when using a function from `LinearAlgebraUtils.h`, stamp its
definition with a `TORCH_API` and that should solve it nikitaved
IvanYashchuk

ghstack-source-id: 8912c32423fe2af9f69552c1773da190146cf0e0
Pull Request resolved: https://github.com/pytorch/pytorch/pull/69824